Prompt to delete when deleting CSS snippet from quick switcher

+++ src/modals/CssSnippetFuzzySuggestModal.ts
@@ -16,6 +16,7 @@ import { detachCssFileLeaves, openView } from "src/obsidian/workspace-helpers";
 import { VIEW_TYPE_CSS } from "src/views/CssEditorView";
 import { CssFile } from "src/CssFile";
 import { handleError } from "src/utils/handle-error";
+import { CssSnippetDeleteConfirmModal } from "./CssSnippetDeleteConfirmModal";
 
 export class CssSnippetFuzzySuggestModal extends FuzzySuggestModal<CssFile> {
 	plugin: CssEditorPlugin;
@@ -238,16 +239,24 @@ export class CssSnippetFuzzySuggestModal extends FuzzySuggestModal<CssFile> {
 					}).catch(handleError);
 				}
 			} else if (evt.key === "Delete") {
-				Promise.all([
-					detachCssFileLeaves(this.app.workspace, item),
-					deleteSnippetFile(this.app, item),
-				])
-					.then(() => {
-						new Notice(`${item.name} was deleted.`);
-					})
-					.catch((err) => {
-						handleError(err, "Failed to delete CSS file.");
-					});
+				if (this.plugin.settings.promptDelete) {
+					new CssSnippetDeleteConfirmModal(
+						this.app,
+						this.plugin,
+						item,
+					).open();
+				} else {
+					Promise.all([
+						detachCssFileLeaves(this.app.workspace, item),
+						deleteSnippetFile(this.app, item),
+					])
+						.then(() => {
+							new Notice(`${item.name} was deleted.`);
+						})
+						.catch((err) => {
+							handleError(err, "Failed to delete CSS file.");
+						});
+				}
 			}
 		} else {
 			const openInNewTab = evt.metaKey;
